Added teacher save part

+++ src/main/java/com/kdpm/schoolTextbookManagement/dto/TeacherDTO.java
@@ -6,7 +6,7 @@ import lombok.*;
 @NoArgsConstructor
 @Data
 public class TeacherDTO {
-    private Long teacherId;
+    private int teacherId;
     private String fullName;
     private String subject;
     private String grade;
